+++ Game/Scripts/Resources.h
@@ -69,8 +69,8 @@ public:
 	void DecreaseDemand(int demand); // Decrease resource demand
 
 	float GetProductionRate(); // Get resource production rate
-	void IncreaseProductionRate(float amountToIncreaseBy); // Increase resource production rate (Example: Due to building upgrades)
-	void DecreaseProductionRate(float amountToDecreaseBy); // Decrease resource production rate (Example: Due to power shortage)
+	void IncreaseProductionRate(int amountToIncreaseBy); // Increase resource production rate (Example: Due to building upgrades)
+	void DecreaseProductionRate(int amountToDecreaseBy); // Decrease resource production rate (Example: Due to power shortage)
 
 	void update(); //Increase/Decrease price if item purchase/sold is detected (based on key input).
 
@@ -83,7 +83,7 @@ private:
 	int itemStock = 0;
 	int basePrice; // Standard price
 	float baseModifier = 5.0;
-	float productionRate = 0.0;
+	int productionRate = 0.0;
 	string itemName, sku, resourceIcon;
 
 	ResourceName resourceName;
